Rename column field names and change data types to appropriate size

+++ database/migrations/2025_12_01_141639_create_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->text('content');
             $table->string('image')->nullable();
-            $table->integer('rating');
+            $table->unsignedSmallInteger('rating');
 
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->foreignId('recipe_id')->constrained()->cascadeOnDelete();
